+++ src/cli/initOpenApiConfig/Enums.ts
@@ -1,0 +1,4 @@
+export enum EOptionType {
+    OPTION = 'OPTION',
+    MULTIOPTION = 'MULTIOPTION'
+}