+++ src/cli/initOpenApiConfig/Types.ts
@@ -1,0 +1,6 @@
+import { EOptionType } from "./Enums";
+
+export type TInitOpenApiConfigParams = {
+    type: EOptionType;
+    openapiConfig?: string;
+}