+++ src/templatesCompiled/cli/openApiConfig.ts
@@ -1,0 +1,20 @@
+// This is an automatically generated file for the hbs template.
+// You don't need to change it, run npm run build:hbs to update it.
+/* istanbul ignore file */
+/* tslint: disable */
+/* eslint: disable */
+// @ts-nocheck
+export default {"1":function(container,depth0,helpers,partials,data) {
+    return "{\n    \"items\": [\n        {\n            \"input\": \"/folder/pogect/openapi/spec_file.yaml\",\n            \"output\": \"./generated/rest-api\"\n        }\n    ],\n    \"httpClient\": \"fetch\",\n    \"sortByRequired\": true\n}\n";
+},"3":function(container,depth0,helpers,partials,data) {
+    return "{\n    \"input\": \"/folder/pogect/openapi/spec_file.yaml\",\n    \"output\": \"./generated/rest-api\",\n    \"httpClient\": \"fetch\",\n    \"sortByRequired\": true\n}\n";
+},"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
+    var stack1, lookupProperty = container.lookupProperty || function(parent, propertyName) {
+        if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
+          return parent[propertyName];
+        }
+        return undefined
+    };
+
+  return ((stack1 = lookupProperty(helpers,"if").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(depth0,"useMultyOptions"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(3, data, 0),"data":data,"loc":{"start":{"line":1,"column":0},"end":{"line":19,"column":7}}})) != null ? stack1 : "");
+},"useData":true}